Fix another getOperatorIndex() bug
I realised just now that I misunderstood the assignment, I was supposed
to *ignore* the proper order of mathematical operations and apply each operation
from a strict left-to-right order. But I had to fix this bug for my ego's sake
anyway. I will now delete all this code and rewrite it properly. It should be
faster doing it the intended way, I think I should be able to do it using two
stacks.